optimized a function

+++ test_project/src/main.f
@@ -1,5 +1,5 @@
 pub function main(): int {
-    int a = 5 as int;
+    int a = 5;
 
     return a;
 }
